pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flax>=0.12.0",
    "iminuit>=2.32.0",
    "ipykernel>=7.0.1",
    "ipython>=9.6.0",
    "jaxlib>=0.8.0",
    "joblib>=1.5.2",
    "markdownify>=1.2.0",
    "matplotlib>=3.10.7",
    "mistune>=3.1.4",
    "nbformat>=5.10.4",
    "optax>=0.2.6",
    "playwright>=1.55.0",
    "polars>=1.34.0",
    "pydantic-ai-slim[openai]>=1.1.0",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "treescope>=0.1.10",
    "wikipedia>=1.4.0",
]

[tool.setuptools]
# disable automatic discovery entirely
packages = []
